Ubuntu Precise task encounters error
On Ubuntu Precise, the error: "mirror does not support the specified release"
occurs. Turns out this is due to a change to how the preseed file is fetched.
Trusty works fine with the current config, so this will change Lucid and
Precise back to the last working version.

This commit follows the decision that the base task should be kept most
up-to-date, meaning that `ubuntu.task` should currently mimic Trusty.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-366