Force Sinatra version above minimum
On earlier versions of Sinatra, the initrd.gz download for e.g. Ubuntu would
hang at a specific percentage because the Content-Length was not being
included. This was fixed in version 1.4.4, and so this gem should be pinned
to this version or later.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-374