Added CoreOS task to Razor.
Allowing for full clusters of CoreOS to be deployed, automatically
configuring the setup with heavy use of Razor's node-metadata function to
enable features like etcd configuration, fleet metadata, formatting of
drives and automatic addition of SSH keys. Since the node-metadata is
applied on all nodes, their value can easily be updated with different
settings on a per-node basis if wanted.