Fix improper abbreviation
The call to the 'node-registered' hook contained an incorrect
assumption that `Razor::Data::Hook` would be aliased to just
`Hook`. This passed in the spec tests, but failed otherwise
yielding an error: "NameError - uninitialized constant
Razor::Command::RegisterNode::Hook". This fully qualifies
that.

For https://tickets.puppetlabs.com/browse/RAZOR-371