Add `--no-content` argument to `create-repo`
Windows ISO extraction is a current known issue.
To subvert this issue, users can manually extract
the ISO and place it in the repo_store directory.
However, previous to this commit, users would
need to enter a stub iso-url to make the
repo's directory in the repo_store. This new
`--no-content` flag will allow the creation of a
repo without an iso-url or url so the repo_store
directory can be manually configured.

For https://tickets.puppetlabs.com/browse/RAZOR-361