Support Protecting New Nodes by Subnet
Adding a configuration option called 'invert_protect_new_nodes_for_subnets' that accepts a list of subnets in CIDR notation (a.a.a.a/bb) for which the 'protect_new_nodes' should be reversed.  This will allow the user to set a default protection value and then override it for specific subnets.