Standardize API argument names from '-' to '_'
In order to standardize the API more, we will use underscores in argument names
rather than dashes. The previous dashes will still be accepted as input, but
the returned values and documentation should be exclusively underscores.

Additionally, if both the dash and underscore arguments are supplied, an
error will be thrown.

For https://tickets.puppetlabs.com/browse/RAZOR-481