Add `http_port` argument for bootstrap over HTTPS
Since booted machines cannot by default use the HTTPS port, we need a way to
generate a bootstrap file that points to the correct HTTP port. That port is
managed outside of Razor, so that requires a new argument, `http_port`, to
the bootstrap endpoint that will be used in the generated bootstrap file.

For https://tickets.puppetlabs.com/browse/RAZOR-487