Add `alias` to Command DSL
This integrates the alias functionality previously used in the `conform!`
method into the DSL itself. This will make aliasing much more standardized.
Aliases will now be on the attribute itself using the keyword `alias` as
either a string or an array of strings.

The framework for this will also automatically create aliases for any
attributes that have underscores in their name to the same name with
dashes, e.g. 'max_count' will automatically have an alias 'max-count'.

Aliases will be reported to clients via the API as
`params.$name.aliases`. The modification from alias to real attribute can take
place at both the client-level (after reading the expected aliases) and at
the server-level (before the `conform!` method is executed).

For https://tickets.puppetlabs.com/browse/RAZOR-481