Fix EL7 broker execution
The broker was not being executed in EL7 builds since the file was not being
marked as executable. It seems like the /etc/rc.d/rc.local file already
existed in EL7 systems, so the permissions were not being set on the file.

For https://tickets.puppetlabs.com/browse/RAZOR-495