RHEL subscription through node metadata
Users will want RHEL systems to be subscribed to RHEL automatically. That can
be done in the kickstart file, but will require some credentials to perform.
There are two methods of doing this:
- Providing the activationkey
- Using auto-attach
Both require a username and password. So, the node metadata used for the
policy is now:
- `rhn_username`
- `rhn_password`
- (optional) `rhn_activationkey`
If the `rhn_activationkey` is omitted, this will use the auto-attach method.
If both `rhn_username` and `rhn_password` are omitted, this step will be
skipped.

The recommended way to add the metadata for this is through the policy's
`node_metadata` property.

For https://tickets.puppetlabs.com/browse/RAZOR-324