Add queue management debugging endpoint
The sequel and hooks queues may need to be cleared for debugging or testing
reasons. This exposes an internal link which will remove those messages. Note
that if this is called, commands may remain in the wrong status, hooks may
not execute, etc.

This will be unlisted, and should not be used for normal operation of Razor.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-537