Allow logging in broker for errors
If the broker hits an issue in its execution, it will be helpful to be able
to reach out to a logging endpoint to store an error. The `log_url` method will
be available to brokers now, taking a `msg` and a `severity`.

For https://tickets.puppetlabs.com/browse/RAZOR-519