Add Windows 2008 R2 task
This includes a task for installing Windows 2008 R2. It includes a few minor
changes to the generic build-razor-winpe.ps1 script.

A few workaround are/were necessary to get this to work.

- The "fastprox.dll" library, which causes issues with WinPE 4.0 + the
  2008R2 installer, is replaced at install-time by the copy included on the
  2008R2 ISO/repo.
- The machine running the build-razor-winpe.ps1 script requires the Windows
  Management Framework (WMF) version 4.0 (which requires that the Windows
  Update service be running in order to install), as well as the usual Windows
  Assessment and Deployment Kit (Windows ADK) version 8.0 or 8.1.

One other modification needed to be made to the generic Windows installer. The
samba share now uses the shell's `hostname` command, rather than the Ruby
URI library's `#hostname` command. The latter retains the domain of the URI,
which is stripped away in samba share workgroups.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-639