Allow reinstall-node with same policy
The reinstall-node command previously required a full re-binding, including a
microkernel boot and potentially binding to a different policy. For
expedited development on tasks, it would be helpful to allow this command to
use the same policy, effectively cutting down the iteration time. This
commit adds a `same_policy` flag which will skip the unbinding step.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-359