Fix razor-server service on EL7
Systemd's .service file was referencing the absent HTTP_PORT variable. This
resulted in an empty string being passed to the service for its http.port
argument. Since an empty string can't be translated to an integer, this error
resulted:

JBAS014688: Wrong type for port. Expected [EXPRESSION, INT] but was STRING

The fix is simply supplying this environment variable, which works here in
the .env file.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-629