Add padding and type checking to sample hostname hook
Hostnames are very often padded, e.g. machine001, machine002, etc. This
would be a helpful feature for the sample hostname hook to have, and so has
been added as a `padding` configuration.

Some of the type checking for both counter and padding were absent, so this
adds that in.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-642