Add `update-broker-configuration` command
Data in the broker configuration may change, for example when a puppet
enterprise master is upgraded to the next version and the `version` config
needs to be changed. Rather than force users to delete and recreate the same
broker with the new datatype, this function treats broker config like hook
config by allowing modification via this new command.