Check to see if Samba share is accessible
If the Samba share with the winpe.wim is not reachable, or the node is unable to create a network connection to it, it would fail, but the script would still complete "successfully".  This would put the node into a bad state where the new OS was not installed, but Razor had it marked as successfully installed.  This simply tests to ensure the node can reach the samba share, and if it is unsuccessful logs the error with razor and exits the script prior to marking the node as installed.