match (but do not bind) when protect_new_nodes is true
Prior to this commit, when protect_new_nodes was set to true, nodes
would not be matched or bound. This commit changes the behavior so
that new nodes will always be matched, even if they are not bound
due to protect_new_nodes being set to true.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-779