Improve error handling for build-razor-winpe.ps1 script
Previously, if the build-razor-winpe.ps1 script hit an error, it would continue
and typically cause more errors. Going back and hunting down the problem in
a terminal is a pain.

This causes errors in commands to halt the script. Additionally, if errors
occur in certain sections, the image that's mounted to the razor-winpe-mount
directory will now be unmounted, making it easier to restart on failure.

For https://tickets.puppetlabs.com/browse/RAZOR-761