Add flag to allow unsigned drivers in WinPE image
For security purposes, the default requirement for drivers to get added to the
WinPE image is that they be signed by a valid authority. Several third-party
drivers exist, and may need to be added to the WinPE image. This commit
includes a flag, "allowunsigned", which, when specified, allows unsigned
drivers to be added to the image. Without this flag supplied, unsigned
drivers in the `extra-drivers` directory will throw an error.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-761