Add "has_macaddress" matcher
Macaddresses are not deterministically reported to Razor from the OS. Facts
like ["in", ["fact", "macaddress"], "00:50:56:a0:bd:24"] may not match
correctly, depending on the order in which interfaces are reported.

The better approach here is to add a "has_macaddress" function which will
search all facts with the macaddress prefix. The function has an
unbounded arity, meaning you can supply as many macaddresses as desired:
["has_macaddress", "mac1", "mac2", ...]

Fixes https://tickets.puppetlabs.com/browse/RAZOR-780