Add /api/collections/config collection
If the user wanted to see which config settings Razor was using, there was
previously no simple way to find out. This commit adds a collection,
`config`, which simply returns the config settings that are in effect.

The collection returns "items", an array which has an object for each
config. The object has key "name" and value "value".

Fixes https://tickets.puppetlabs.com/browse/RAZOR-531