`no_replace` conflicts should result in an error
When the user runs either `modify-node-metadata` or `update-node-metadata`
using the `no_replace` flag, the result is that the operation fails silently
for any key(s) that already exist. This is not the behavior documented for
these commands, which state that the command will fail if this happens.

In the case of `modify-node-metadata`, a new flag `force` will be added so that
batch operations can still succeed. This will allow feature parity with the
existing behavior by supplying a value of `true`.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-940