Add `update-policy-node-metadata` command
Users may need to modify the node metadata that gets assigned when a node binds
to a policy. There is currently no command to perform this, so the user must
work around the issue by either deleting the policy or creating a specialized
hook to achieve this.

This commit creates a new command, `update-policy-node-metadata`, to handle
this case.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-672