Node Metadata updates (#361)
* `no_replace` conflicts should result in an error

When the user runs either `modify-node-metadata` or `update-node-metadata`
using the `no_replace` flag, the result is that the operation fails silently
for any key(s) that already exist. This is not the behavior documented for
these commands, which state that the command will fail if this happens.

In the case of `modify-node-metadata`, a new flag `force` will be added so that
batch operations can still succeed. This will allow feature parity with the
existing behavior by supplying a value of `true`.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-940

* Add `update-policy-node-metadata` command

Users may need to modify the node metadata that gets assigned when a node binds
to a policy. There is currently no command to perform this, so the user must
work around the issue by either deleting the policy or creating a specialized
hook to achieve this.

This commit creates a new command, `update-policy-node-metadata`, to handle
this case.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-672