Add broker for Puppet 4 (#364)
Now that Puppet 3 is approaching EOL, we should include a stock broker for
Puppet 4. Since the ideal name for the broker is `puppet`, the right approach
is to rename existing `puppet` broker to `legacy-puppet` and add a new `puppet`
broker for the new AIO broker.

If the user wants to change a policy to point to the old `puppet` broker, the
new `update-policy-broker` command will facilitate this. That will allow a
seamless migration.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-934