Allow `mac` in `set-node-hw-info`
One confusing aspect about how Razor identifies nodes is that the mac
addresses are passed in as `net0`, `net1`, etc., but are represented in Razor
as `mac`, an array of all of these. To mitigate that confusion, the
`set-node-hw-info` command will now accept the `mac` attribute directly.

Additionally, when only `net0` was supplied, it was not being properly
converted into `mac` before the validation checks. This meant an error was
thrown saying that `mac` is a required attribute.

Now, the user can supply `netX`, `mac`, or both methods.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-939