Store all `hw_info` when nodes are initially registered
Whenever a node is initially added to the system, the only `hw_info` values
that are recorded are the values that are used to match the node. This can be
a problem if the value used to match nodes changes, because there will be no
previous value to compare to in order to find the same node in the system.

To solve this, we will store all hardware info values that are passed by iPXE.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-958