Allow deletion of properties not in schema for on-disk types
When a config property from a broker-type on disk gets removed, any objects
which have those properties defined will fail the schema validation. If the user
tried to convert these brokers to a valid state, the
`update-broker-configuration` command would fail to clear the property.

This fixes that by allowing modification of objects that fail the schema in
both `update-broker-configuration` and `update-hook-configuration`.