Perform locale negotiation
Razor already has all user-facing messages going through the workflow to be
translated, but the locale (sent through request headers) is not being taken
into account when sending the messages to the user.

This adds the final step in the process to allow new translations to be
visible to the user by adding the `gettext-setup` gem to the project and
sending the accept language header to be interpreted by that gem.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-651