(maint) do not require config.yaml (#113)
* (maint) do not require config.yaml

Prior to this commit, razor was requiring config.yaml be present
even though it is not needed and is not currently laid down by
the module. This commit eliminates the requirement for that file.

* (PE-10198) Return values has on config init

The config init used to return the @values instance variable, due to
the @values.merge being the last call made in the method. This commit
returns that behavior in hopes of fixing the issue of the config being
missing on startup.