Fix `repo_file` task function for repos with remote source
The `repo_file` function, accessible by tasks, currently only looks for local
files in the repo directory. This works fine when the repo is created with the
`iso_url` argument, but not when it's using a remote source via the `url`
argument.

The fix is for this function is to download from a remote source when `url` is
set, otherwise use the current behavior of using a local source's contents.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-1022