Add ESXi 6 support
This adds stock task support for ESXi 6. This is a nominal change which will
allow future expansion as we support more versions. Following the standard of
other tasks, the `vmware_esxi` task will always reference the latest version of
the OS that we support. This commit also adds two new tasks: `vmware_esxi/5.5`
and `vmware_esxi/6`. These two tasks are functionally identical, but if we add
a later task that requires changes to the task's files, the 5.5 and 6 tasks will
remain functionally identical to how they are now.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-967