Add Ubuntu Xenial task
This adds a task for Ubuntu Xenial. Since that is the latest version of Ubuntu
that we support, that also means the generic `ubundu` task should point to this.
An `ubuntu/xenial` task now exists, which will work for Xenial even when new
versions are added.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-854

Update to 2017.3.2