Update to JRuby 9000 and Torquebox 3.2.0
To stay ahead of updates to our dependencies and ensure our
gems get proper security fixes, we should move to JRuby 9.1.5.0. This
also allows us to update to the latest version of Torquebox: 3.2.0.

This commit also addresses a few things that have changed in our
dependencies.

On the sequel gem side, serialization now happens prior to validation.
That has two implications: we need to re-serialize objects if they
change in their validation, and serialization needs to be robust enough
that it will not fail prior to the object being validated.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-973