Skip certain tests if the user is running as root
Previously, a test in spec/broker_type_spec.rb and another test in
spec/data/repo_spec.rb set the permissions of a file to "0000" with
the intent that the codebase would throw an exception if that file
was not readable. When running as root, however, everything is
readable so these tests failed. This commit skips these tests when
running as root, because they don't make sense for that specific
case.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-1050