Pass-thru if stdin.write fails with Errno::EPIPE in exec_script
Previously, the tests in spec/data/hook_spec.rb would transiently
fail at the stdin.write step in exec_script due to an Errno::EPIPE
exception. Errno::EPIPE is thrown because of a race condition: the
process running the script exited before stdin.write was called. This
happened because the test scripts finished very quickly. From a user's
perspective, this would mean that the exec_script task has a chance
of failure if the executed script finishes quickly, which is undesirable.

This commit fixes things so that when stdin.write fails with an Errno::EPIPE
exception, the code just passes through. This should be OK, because if something
goes wrong with the process itself, that should be captured in the process' success
status queried at the end of exec_script.

The code also cleans up exec_script to improve its readability.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-1051