Add a 'depth' parameter to the /api/collections/<type> endpoint
Previously, the /api/collections/<type> endpoint would return a list
of object references to the items in the collection. For people who
wanted to see the complete details of each item in the collection,
they would have needed to iterate over the item references and make
a total of N requests to their corresponding API endpoints, where N
is the size of the collection. This is annoying when writing scripts,
and it does not scale well for large collections.

This commit introduces an optional depth parameter that can be one
of two possible values:
  (1) 0 -- Returns a list of object references for the items
  in the collection, which is the default behavior.

  (2) 1 -- Returns a detailed list of the items in the collection,
  meaning that for each entry <item> in the collection, the response body
  of the endpoint /api/collections/<type>/<item> is returned.

The default depth is 0 for backwards compatibility.

Linked to https://tickets.puppetlabs.com/browse/RAZOR-433