(maint) Consolidate config files
This commit removes the razor-server.env file (used only by the systemd service
file) and moves all its variable assignments to the razor-server.sysconfig,
which is used by both the sysV init script and the systemd service file. This
way, it will be clearer and more consistent how the service receives arguments
and will allow these args to be managed by the puppetlabs-pe_razor module.